requirements.txt
# ── Temporal SDK ──────────────────────────────────────────────
temporalio>=1.9.0

# ── Configuration ────────────────────────────────────────────
python-dotenv>=1.0.0
boto3>=1.34.0

# ── Microsoft Graph SDK ──────────────────────────────────────
azure-identity>=1.15.0
msgraph-sdk>=1.2.0

# ── Validation / Models ──────────────────────────────────────
pydantic>=2.0

# ── Web framework (webhook trigger) ─────────────────────────
fastapi>=0.115.0
uvicorn>=0.34.0
httpx>=0.27.0

# ── Testing ──────────────────────────────────────────────────
pytest>=8.0
pytest-mock>=3.14.0
pytest-asyncio>=0.24.0
